working version
Running this project creates public/private keys in Wallet Import Format (WIF), encodes them as QR codes and creates a PDF with the QR codes and the WIF as text.  At this point it's customized to my included background image, but a future version could be created to add your own image.